# agri-backend/openapi-spec-preview.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AgriMart API",
    "version": "1.0.0",
    "description": "Production REST API for AgriMart, Zambia's agricultural marketplace connecting farmers and buyers. Backed by PostgreSQL (Supabase). All endpoints return JSON. Authenticated endpoints require `Authorization: Bearer <token>`, where the token is issued by `POST /api/auth/login` and is valid for 7 days.\n\n### Map coordinates\nProducts may carry a `latitude`/`longitude` pin (decimal degrees, WGS84 — the same format returned by the browser Geolocation API and by map libraries like Leaflet/Google Maps). These are optional on every product write, and are what a frontend uses to plot farmer locations on a map. There is currently no separate location field on a user/profile — only on individual products.\n\n### Field casing\nMost resources (users, products) are returned in camelCase. The one exception is `/api/notifications`, which returns raw database column names in snake_case (`email_notifications`, `push_notifications`, `sms_notifications`, `price_predictions_enabled`) — this is intentional, matching the actual response shape, not a documentation error.",
    "contact": {
      "name": "AgriMart"
    }
  },
  "servers": [
    {
      "url": "https://online-agriculture-market-place.onrender.com",
      "description": "production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Obtain a token from POST /api/auth/login, valid for 7 days."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "stack": {
            "type": "string",
            "description": "Only present when NODE_ENV=development."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Availability": {
        "type": "string",
        "enum": [
          "available",
          "limited",
          "sold_out",
          "hidden"
        ],
        "default": "available",
        "description": "`hidden` listings are excluded from every public read (list, search, compare, saved, and even the owner viewing their own supplier filter)."
      },
      "CategoryId": {
        "type": "string",
        "enum": [
          "grains",
          "vegetables",
          "fruits",
          "legumes",
          "tubers"
        ],
        "example": "vegetables",
        "description": "Matches the seeded `categories` table."
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "White Maize"
          },
          "categoryId": {
            "$ref": "#/components/schemas/CategoryId"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 175
          },
          "unit": {
            "type": "string",
            "example": "50kg bag"
          },
          "quantity": {
            "type": "integer",
            "example": 200
          },
          "description": {
            "type": "string",
            "example": "Freshly harvested, well dried."
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://.../products/xyz.jpg"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid",
            "description": "Matches a User.id — join against GET /api/users/:id for supplier details."
          },
          "location": {
            "type": "string",
            "example": "Chongwe"
          },
          "province": {
            "type": "string",
            "example": "Lusaka"
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true,
            "minimum": -90,
            "maximum": 90,
            "example": -15.4067,
            "description": "Map pin latitude in decimal degrees (WGS84). Null if this listing has not been pinned on a map yet."
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true,
            "minimum": -180,
            "maximum": 180,
            "example": 28.2871,
            "description": "Map pin longitude in decimal degrees (WGS84). Pairs with `latitude`."
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "createdAt": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "CreateProductInput": {
        "type": "object",
        "description": "Sent as multipart/form-data (to allow an optional image file alongside the fields). All values below arrive as form fields; `image` is the file part.",
        "required": [
          "name",
          "categoryId",
          "price",
          "unit",
          "quantity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "White Maize"
          },
          "categoryId": {
            "$ref": "#/components/schemas/CategoryId"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "example": 175
          },
          "unit": {
            "type": "string",
            "example": "50kg bag"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "example": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "example": "Freshly harvested."
          },
          "location": {
            "type": "string",
            "maxLength": 100,
            "example": "Chongwe"
          },
          "province": {
            "type": "string",
            "maxLength": 100,
            "example": "Lusaka"
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "example": -15.4067,
            "description": "Optional map pin latitude — drop a pin for this listing."
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "example": 28.2871,
            "description": "Optional map pin longitude — pairs with `latitude`."
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Optional. jpeg/jpg/png/gif/webp, up to 5MB."
          }
        }
      },
      "UpdateProductInput": {
        "type": "object",
        "description": "multipart/form-data, same shape as create but every field is optional — send only what changed. At least one field is required.",
        "properties": {
          "name": {
            "type": "string"
          },
          "categoryId": {
            "$ref": "#/components/schemas/CategoryId"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "unit": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "image": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "ComparableProduct": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "CompareResponse": {
        "type": "object",
        "properties": {
          "reference": {
            "$ref": "#/components/schemas/Product"
          },
          "comparable": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComparableProduct"
            },
            "description": "Up to 20 other listings in the same category, cheapest first."
          }
        }
      },
      "PricePrediction": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "predictedPrice": {
            "type": "number"
          },
          "confidence": {
            "type": "integer",
            "description": "Percent, 70–100."
          }
        }
      },
      "PredictionsResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "currentPrice": {
            "type": "number"
          },
          "predictions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PricePrediction"
            }
          },
          "model": {
            "type": "string",
            "example": "dummy",
            "description": "Currently a placeholder model — not real forecasting yet."
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Request body/query failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed, or expired bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "No resource matches this id (or it belongs to someone else).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Database is not ready yet.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "List marketplace products",
        "description": "Excludes `hidden` listings. Ordered newest first.",
        "tags": [
          "Products"
        ],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "$ref": "#/components/schemas/CategoryId"
            },
            "description": "Filter by category id."
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive match against name, location, or description.",
            "example": "tomato"
          },
          {
            "in": "query",
            "name": "supplier",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter to one supplier's listings (their user id)."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "summary": "Create a product listing",
        "description": "Only farmers should list products (not enforced by role at this route, but intended usage). Include `latitude`/`longitude` to pin this listing's location on a map.\n",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/products/compare": {
      "get": {
        "summary": "Compare one product's price against similar listings",
        "description": "Returns the reference product plus up to 20 other listings in the same category (cheapest first, excluding hidden listings).\n",
        "tags": [
          "Products"
        ],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comparison result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing `productId`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/products/predictions": {
      "get": {
        "summary": "Get price trend predictions for a product",
        "description": "Requires the caller's `price_predictions_enabled` notification setting to be true (see PUT /api/notifications). The model is currently a placeholder (`model: \"dummy\"`), not real forecasting.\n",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Predicted prices.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing `productId`, or `days` is not a positive integer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Price predictions are disabled for this account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get a single product",
        "tags": [
          "Products"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Update a product listing you own",
        "description": "Only the listing's owner (`supplierId` matching the token) can update it. Send only the fields that changed.\n",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found, or it belongs to a different account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a product listing you own",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found, or it belongs to a different account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}/mark-sold-out": {
      "post": {
        "summary": "Mark a product you own as sold out",
        "description": "Sets `availability` to `sold_out` and `quantity` to 0.",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found, or it belongs to a different account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}